Add redux-thunk for async data fetching
configure store as a reusable file.
